requirements.txt
faker
pandas
fastapi
uvicorn
streamlit
chromadb
langchain
langchain-huggingface
sentence-transformers
langchain_google_genai
sqlglot
sqlalchemy
python-dotenv
langgraph
langchain-chroma
tqdm
pydantic
requests
aiohttp
langchain-community
langchain-core
